minor up

+++ CrossSection/run_fitX.sh
@@ -7,7 +7,7 @@ centmax=90
 ymin=0
 ymax=1.6
 
-counts=(1)
+counts=(0 1 2 3)
 optcuts=(
     "&& BDTF > 0.3 && (Bmass-3.096916-Btktkmass) < 0.12"
     "&& BDT > 0.07 && (Bmass-3.096916-Btktkmass) < 0.12"
@@ -40,39 +40,41 @@ RUN_CONVERTTNP=${2:-0}
 RUN_FITHIST=${3:-0}
 RUN_DRAWHIST=${4:-0}
 
+tmp=$(date +%y%m%d%H%M%S)
+
 ##
-g++ getfname.cc $(root-config --libs --cflags) -g -o getfname.exe || exit 1
-kinematic=$(./getfname.exe $ptmin $ptmax $centmin $centmax $ymin $ymax)
-rm getfname.exe
+g++ getfname.cc $(root-config --libs --cflags) -g -o getfname_${tmp}.exe || exit 1
+kinematic=$(./getfname_${tmp}.exe $ptmin $ptmax $centmin $centmax $ymin $ymax)
+rm getfname_${tmp}.exe
 echo -e "\e[32;1mcompiling...\e[0m"
 cp ../tnp/tnpcc.h tnpcc_tmp.h
 ptbins='float ptbins[] = {'$ptmin', '$ptmax'};'
 sed -i "s/__PTBIN_INPUT__/$ptbins/g" tnpcc_tmp.h
 [[ $RUN_SAVEHIST -eq 1 || $# -eq 0 ]] && {
     echo " -- "fitX_flatten.C
-    g++ fitX_flatten.C $(root-config --libs --cflags) -g -o fitX_flatten.exe || exit 1 
+    g++ fitX_flatten.C $(root-config --libs --cflags) -g -o fitX_flatten_${tmp}.exe || exit 1 
     echo " -- "fitX_savehist.C
-    g++ fitX_savehist.C $(root-config --libs --cflags) -lRooFit -lRooFitCore -g -o fitX_savehist.exe || exit 1 
+    g++ fitX_savehist.C $(root-config --libs --cflags) -lRooFit -lRooFitCore -g -o fitX_savehist_${tmp}.exe || exit 1 
 }
 [[ $RUN_FITHIST -eq 1 || $# -eq 0 ]] && {
     echo " -- "draw_tnp.cc
-    g++ draw_tnp.cc $(root-config --libs --cflags) -g -o draw_tnp.exe || exit 1 
+    g++ draw_tnp.cc $(root-config --libs --cflags) -g -o draw_tnp_${tmp}.exe || exit 1 
     echo " -- "fitX_fithist.C
-    g++ fitX_fithist.C $(root-config --libs --cflags) -lRooFit -lRooFitCore -g -o fitX_fithist.exe || exit 1
+    g++ fitX_fithist.C $(root-config --libs --cflags) -lRooFit -lRooFitCore -g -o fitX_fithist_${tmp}.exe || exit 1
 }
 [[ $RUN_DRAWHIST -eq 1 || $# -eq 0 ]] && {
     echo " -- "fitX_drawhist.C
-    g++ fitX_drawhist.C $(root-config --libs --cflags) -g -o fitX_drawhist.exe || exit 1
+    g++ fitX_drawhist.C $(root-config --libs --cflags) -g -o fitX_drawhist_${tmp}.exe || exit 1
 }
 
 ##
 [[ $RUN_SAVEHIST -eq 1 ]] && {
     flatinput=${input%%.root}_flatten.root
-    [[ -f $flatinput ]] || ./fitX_flatten.exe $input $flatinput
+    [[ -f $flatinput ]] || ./fitX_flatten_${tmp}.exe $input $flatinput
     flatinputmc_a_prompt=${inputmc_a_prompt%%.root}_flatten.root
-    [[ -f $flatinputmc_a_prompt ]] || ./fitX_flatten.exe $inputmc_a_prompt $flatinputmc_a_prompt
+    [[ -f $flatinputmc_a_prompt ]] || ./fitX_flatten_${tmp}.exe $inputmc_a_prompt $flatinputmc_a_prompt
     flatinputmc_b_prompt=${inputmc_b_prompt%%.root}_flatten.root
-    [[ -f $flatinputmc_b_prompt ]] || ./fitX_flatten.exe $inputmc_b_prompt $flatinputmc_b_prompt
+    [[ -f $flatinputmc_b_prompt ]] || ./fitX_flatten_${tmp}.exe $inputmc_b_prompt $flatinputmc_b_prompt
 }
 
 ##
@@ -82,11 +84,12 @@ do
     cut="HLT_HIL3Mu0NHitQ10_L2Mu0_MAXdR3p5_M1to5_v1 && pprimaryVertexFilter && phfCoincFilter2Th4 && pclusterCompatibilityFilter"
     cut="$cut && mvapref ${optcuts[count]}"
     cutgen="1"
-    cp ../tnp/tnp_converter.cc tnp_converter_tmp.cc
-    sed -i "s/__CUTINPUT__/${optcutntuples[$count]}/g" tnp_converter_tmp.cc
+    cp ../tnp/tnp_converter.cc tnp_converter_tmp_${tmp}.cc
+    sed -i "s/__CUTINPUT__/${optcutntuples[$count]}/g" tnp_converter_tmp_${tmp}.cc
     [[ $RUN_CONVERTTNP -eq 1 || $# -eq 0 ]] && {
-        g++ tnp_converter_tmp.cc $(root-config --libs --cflags) -g -o tnp_converter.exe || continue
+        g++ tnp_converter_tmp_${tmp}.cc $(root-config --libs --cflags) -g -o tnp_converter_${tmp}.exe || { rm tnp_converter_tmp_${tmp}.cc ; continue ; }
     }
+    rm tnp_converter_tmp_${tmp}.cc
 
     name=trainX_20190808ptdep_sideband_tktk0p2_15p0_50p0_0-10-1-2-9_${tags[count]}
     rootdir=rootfiles/$name$kinematic/
@@ -95,33 +98,33 @@ do
     echo -e "----------------------------------------"
 
     [[ $RUN_SAVEHIST -eq 1 ]] && {
-        ./fitX_savehist.exe $input $inputmc_a_prompt $inputmc_b_prompt $inputmc_a_nonprompt $inputmc_b_nonprompt "$cut" "$cutgen" "$name" $ptmin $ptmax $centmin $centmax $ymin $ymax
+        ./fitX_savehist_${tmp}.exe $input $inputmc_a_prompt $inputmc_b_prompt $inputmc_a_nonprompt $inputmc_b_nonprompt "$cut" "$cutgen" "$name" $ptmin $ptmax $centmin $centmax $ymin $ymax
     }
     [[ $RUN_CONVERTTNP -eq 1 ]] && {
-        ./tnp_converter.exe $inputmc_a_prompt $name "_a" $ptmin $ptmax $centmin $centmax $ymin $ymax
-        ./tnp_converter.exe $inputmc_b_prompt $name "_b" $ptmin $ptmax $centmin $centmax $ymin $ymax
+        ./tnp_converter_${tmp}.exe $inputmc_a_prompt $name "_a" $ptmin $ptmax $centmin $centmax $ymin $ymax
+        ./tnp_converter_${tmp}.exe $inputmc_b_prompt $name "_b" $ptmin $ptmax $centmin $centmax $ymin $ymax
+        rm tnp_converter_${tmp}.exe
     }
+
     [[ $RUN_FITHIST -eq 1 ]] && {
-        ./draw_tnp.exe "$rootdir/tnp_a.root" $name "_a"
-        ./draw_tnp.exe "$rootdir/tnp_b.root" $name "_b"
-        ./fitX_fithist.exe "$rootdir/fitX_savehist.root" $name
+        ./draw_tnp_${tmp}.exe "$rootdir/tnp_a.root" $name "_a"
+        ./draw_tnp_${tmp}.exe "$rootdir/tnp_b.root" $name "_b"
+        ./fitX_fithist_${tmp}.exe "$rootdir/fitX_savehist.root" $name
     }
-    [[ $RUN_DRAWHIST -eq 1 ]] && ./fitX_drawhist.exe "$rootdir/fitX_fithist.root" $name
+    [[ $RUN_DRAWHIST -eq 1 ]] && ./fitX_drawhist_${tmp}.exe "$rootdir/fitX_fithist.root" $name
 
     # syst
     # echo fitX_pdfvar.C
-    # g++ fitX_pdfvar.C $(root-config --libs --cflags) -lRooFit -lRooFitCore -g -o fitX_pdfvar.exe || exit 1
-    # [[ ${5:-0} -eq 1 ]] && ./fitX_pdfvar.exe "$outputdir/fitX_savehist" $name
+    # g++ fitX_pdfvar.C $(root-config --libs --cflags) -lRooFit -lRooFitCore -g -o fitX_pdfvar_${tmp}.exe || exit 1
+    # [[ ${5:-0} -eq 1 ]] && ./fitX_pdfvar_${tmp}.exe "$outputdir/fitX_savehist" $name
 
-    [[ -f tnp_converter.exe ]] && rm tnp_converter.exe
-    rm tnp_converter_tmp.cc
 done
 
-[[ -f fitX_drawhist.exe ]] && rm fitX_drawhist.exe
-[[ -f fitX_fithist.exe ]] && rm fitX_fithist.exe
-[[ -f fitX_savehist.exe ]] && rm fitX_savehist.exe
-[[ -f fitX_flatten.exe ]] && rm fitX_flatten.exe
-[[ -f draw_tnp.exe ]] && rm draw_tnp.exe
-rm tnpcc_tmp.h
+[[ -f fitX_drawhist_${tmp}.exe ]] && rm fitX_drawhist_${tmp}.exe
+[[ -f fitX_fithist_${tmp}.exe ]] && rm fitX_fithist_${tmp}.exe
+[[ -f fitX_savehist_${tmp}.exe ]] && rm fitX_savehist_${tmp}.exe
+[[ -f fitX_flatten_${tmp}.exe ]] && rm fitX_flatten_${tmp}.exe
+[[ -f draw_tnp_${tmp}.exe ]] && rm draw_tnp_${tmp}.exe
+[[ -f tnpcc_tmp.h ]] && rm tnpcc_tmp.h
 
-# rm fitX_pdfvar.exe
+# rm fitX_pdfvar_${tmp}.exe
